TruthLens/backend/requirements.txt
jupyter==1.0.0
gradio
Pillow==10.4.0
facenet-pytorch==2.5.2
opencv-python==4.7.0.72
grad-cam==1.4.6
html2image~=2.0.4.3
weasyprint
fastapi
requests
motor
pymongo
pymongo[srv]
pydantic
uvicorn
pytube
starlette
python-dotenv~=0.16.0
web3
yt-dlp
uvicorn
numpy==1.26.4
eth-brownie==1.20.6